NIG webservice
Create a .NET (Framework, not Core) webservice which accepts a string and returns a string. It should:

expect to receive a string which is a day of the week (Monday, Tuesday etc).
return the date which is the next day of that name, based on the current UTC date/time.
never return the current day, even if it matches.